Fixed some bugs, added the ability to change the progress of a message, added websocket automatic reconnecting.

+++ APIProgressTracker/JSONObjects/Message.cs
@@ -11,7 +11,7 @@ namespace APIProgressTracker.JSONObjects
         public string ID { get; set; }
         public string Title { get; set; }
         public string Contents { get; set; }
-        public short Progress { get; set; }
+        public int Progress { get; set; }
         public Author Author { get; set; }
         public DateTime LastModified { get; set; }
     }
